+++ Assets/Dust/CHANGELOG.md
@@ -11,7 +11,9 @@
 - Removed prefix "Du" for all Factory & FactoryMachines
 - Removed prefix "Du" for all Fields
 - Removed prefix "Du" for all Gizmos
-- Rename prefix "Du" to "On" for all Events
+- Removed prefix "Du" for all Helpers
+- Renamed prefix "Du" to "On" for all Events
+- Renamed class "DuDebug" to "Debugger"
 
 
 ________________________________________________________________________________________________________________________